refactor(resource): move functionality from component into helper files
