+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-25T13:31:50.345Z\n"
-"PO-Revision-Date: 2019-02-25T13:31:50.345Z\n"
+"POT-Creation-Date: 2019-02-25T14:07:50.878Z\n"
+"PO-Revision-Date: 2019-02-25T14:07:50.878Z\n"
 
 msgid "Cancel"
 msgstr ""
