+++ src/utils/api.js
@@ -11,7 +11,6 @@ import {
     STANDARD_REPORTS_ENDPOINT,
     DATA_SET_REPORTS_ENDPOINT,
     DATA_SET_DIMENSIONS_ENDPOINT,
-    REPORTING_RATE_SUMMARY_ENDPOINT,
     RESOURCE_ENDPOINT,
     DATA_DIMENSION_SUFFIXES,
     postDataSetReportCommentUrl,
